quick and dirty or

+++ src/base16.h
@@ -99,7 +99,7 @@ public:
         const int t16(16);
         int remainder;
         while (n != 0) {
-            remainder = n % t16;
+            remainder = static_cast<int>(n % t16);
             n /= t16;
             ss << kValidChars[remainder];
         }
